+++ roseau/load_flow/models/switches.py
@@ -55,6 +55,7 @@ class Switch(AbstractBranch[CySwitch]):
         super().__init__(id=id, phases1=phases, phases2=phases, bus1=bus1, bus2=bus2, geometry=geometry)
         self._check_elements()
         self._check_loop()
+        self._check_same_voltage_level()
         self._cy_element = CySwitch(self._n1)
         self._cy_connect()