+++ roseau/load_flow/models/tests/test_branches.py
@@ -1,4 +1,7 @@
 import numpy as np
+import pytest
+
+from roseau.load_flow.models import Bus, Line, LineParameters, Switch
 
 
 def test_powers_equal(network_with_results):
@@ -9,3 +12,21 @@ def test_powers_equal(network_with_results):
     assert np.allclose(sum(powers1), -sum(vs.res_powers))
     assert np.allclose(sum(powers2), -sum(pl.res_powers))
     assert np.allclose(powers1 + powers2, line.res_power_losses)
+
+
+def test_different_voltage_levels():
+    bus1 = Bus(id="bus1", phases="abc", nominal_voltage=240)
+    bus2 = Bus(id="bus2", phases="abc", nominal_voltage=240)
+    bus3 = Bus(id="bus3", phases="abc")
+    bus4 = Bus(id="bus4", phases="abc", nominal_voltage=400)
+    lp = LineParameters(id="lp", z_line=np.eye(3, dtype=complex))
+    Line(id="ln good", bus1=bus1, bus2=bus2, parameters=lp, length=0.1)  # OK
+    Line(id="ln good2", bus1=bus1, bus2=bus3, parameters=lp, length=0.1)  # OK
+    with pytest.warns(UserWarning, match=r"Line 'ln bad' connects buses with different nominal voltages: 240 and 400."):
+        Line(id="ln bad", bus1=bus1, bus2=bus4, parameters=lp, length=0.1)
+    Switch(id="sw good", bus1=bus1, bus2=bus2)  # OK
+    Switch(id="sw good2", bus1=bus1, bus2=bus3)  # OK
+    with pytest.warns(
+        UserWarning, match=r"Switch 'sw bad' connects buses with different nominal voltages: 240 and 400."
+    ):
+        Switch(id="sw bad", bus1=bus1, bus2=bus4)